backend/auth_utils.py
from datetime import datetime,timedelta

from jose import JWTError,jwt
import os
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY", "fallback-secret-key-change-in-production")

ALGORITHM = "HS256"


ACCESS_TOKEN_EXPIRE_MINUTES = 60




def create_access_token(data:dict):

    to_encode = data.copy()

    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode.update({"exp":expire})

    return jwt.encode(to_encode,SECRET_KEY,algorithm=ALGORITHM)



def decode_access_token(token:str):

    try:

        return jwt.decode(token,SECRET_KEY,algorithms=[ALGORITHM])
    
    except JWTError:

        return None