backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.3
openai==1.12.0
email-validator==2.1.0
